frontend-plugin-api: move extension instance creation to frontend-app-api
